examples/advanced/03_web_frameworks/django_project/requirements.txt
# Django 프로젝트 요구사항
pyhub-llm[all]
django>=4.0
djangorestframework
django-cors-headers
pillow  # 이미지 처리
